requirements.txt
pandas
pandas-datareader
pandas-gbq
pandas-stubs
numpy
sklearn-pandas
scikit-learn
seaborn
flask
imblearn
openpyxl